feat : 마이페이지에서 메뉴판 삭제 기능 추가
