+++ frontend/components/AdminMenu/AdminMenu.tsx
@@ -30,7 +30,7 @@ const AdminMenu = ({ menuList }: Props) => {
       <S.MenuContent>
         {menuList &&
           menuList.map(({ id, name }) => (
-            <MenuCard key={id} id={id} name={name} />
+            <MenuCard key={id} menuId={id} menuName={name} />
           ))}
       </S.MenuContent>
       {modalOpen && <MenuModal setModalOpen={setModalOpen} type="메뉴판" />}
